backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
python-multipart
sqlalchemy
requests
python-dotenv
